Start MonkeyKitSocketService when gcm receives msg
Properly manage realm instances. Open them onStart and close them onStop. Avoid
global realm instances.
DatabaseHandler loads all messages when app starts. not just the ones sent by
the current user

Always return a MonkeyBinder on the onBind method. this ensures that the
delegate is always set correctly.

Update lastTimeSynced only when a message or a messageBatch is received.
updating when sync is sent does not guarantee that the response was received.

If the service is manually started dont stop it after the batch is received.
Instead start ServiceTimeoutTask.

ServiceTimeoutTask now has a 10 seconds timeout. This timeout is enforced with a
static variable, so that no matter how many tasks have been spawned, there
should always be a 10 seconds idle time before stopping the service.

MonkeyKitSocketService has new abstract methods. openDatabase() and
closeDatabase(). These are called when the service initializes and gets
destroyed so that it is easier for the developer to manage his/her database.

Renamed executeInDelegate method to processMessageFromHandler. This method now
does a little more than just forwarding stuff to the delegates.

MainActivity now sends messages to a fixed monkeyID. This is useful for testing
receiving messages from other phones.