Create notifications with MyGcmListenerService
MonkeyGcmListenerService now has 2 extra abstract methods, createNotification
and createLocalizedNotification. The developer can override these methods to
post notifications to the device when a push message is received via gcm and the
user is not using the app.

Add a new ServiceStatus: bound. with this status is easier for the
MonkeyGcmListenerService to know if the service is running AND has a bound
delegate.