new delegate callback: onConnectionRefused
When the developer used wrong credentials, MonkeyKit would enter an endless loop
of reconnecting. Since the server would never accept the login requests with
such credentials, this loop would continue until the OS kills the process.

To fix this, when the FAILLOGGINMSG is received in the DarkStarSocketClient, the
connection thread is interrupted and the socketStatus is set as UNAUTHORIZED,
which is a new socketStatus, and the message is passed to the Application using
the new callback: onConnection Refused. When the socket is unathorized, it
should not retry connections anymore. The developer should use this callback to
inform the user that the service is unavailable.