refactor storeMessage() into storeReceivedMessage()
The MonkeyKitSocketService storeMessage has been renamed storeReceivedMessage().
This is now exclusively used for received messages.
Additionally, MKDelegateActivity has a new abstract method: storeSentMessage().
This one is exclusively used for messages that the user sends.
The service no longer store's the user's messages. That's the delegate's job.
The distinction is necessary because sometimes the delegate may need to persist
messages while the service is not available. This is a problem we encountered
while trying to send photos