Upload/Download file with intent service
MonkeyFileService is a new service that uploads and downloads files via HTTP.
It does not need MonkeyKitSocketService to be active to finish networking.
It also updates the database once it finishes.FileManager class has been
deprecated and deleted.

The delegate interface has a new method: onFileDownloadFinished. This replaces
the MonkeyHttpResponse object.

The developer must subclass this service and override the function that can
update the database: onFileUploadFinished and onFileUploadFinished
.
MonkeyKitSocketService also has a new abstract method that must be overriden so
that it can know the class name of the developer's file service.

MOKMessage was rewritten in Kotlin and it now has a constructor from Intent and
a method to convert to intent.