Improved reconnection with bad network
Watchdog now uses a exponential periodic function for the timeout. This is to
prevent pontentially infinitely large timouts.

when the handler sends a onSocketConnected message to the service,
resendPendingMessages() is executed. this fixes the issue of text messages not
being resent when the socket reconnects.

SampleApp's DatabaseHandler now throws an exception when it tries to store a
message that already exists. This is done to prevent the runnable from executing
onMessageReceived() with messages that already exist. SaveMessage renamed into
SaveIncomingMessage().

MsgSenderService class was deleted.

better documentation on AsyncAESInitializer