Fix duplicate handler issue
MOKMessageHandler is no longer initialized using kotlin's lazy. it now has a
method clearServiceReference() so that the WeakReference can be released. When
the service is destroyed the reference is cleared. When service is bound a new
MOKMessageHandler instance is created.

startSocketConnection will not do anything if the service status is equal to
'Dead' this prevents reinitializing the socket after destroying the service.