Retain data with fragment on configuration change
ChatDataFragment class stores MainActivity's data onConfigurationChange.
reorganize MainActivity's interface methods.

MessageLoader class was modified to be easier to retain. Messages are now loaded
before instantiating the chat fragment. MessageLoader has a new inner class
MessageCounter that keeps track of the number of loaded messages. The
RealmChangeListener was moved to a inner class NewMessagesListener.
In addition to the loadNewPage method, there is also a loadFirstPage, which
instantiates the Chat fragment after loading the first batch of messages.