Send open/close conversation
MKDelegateActivity has a new property: activeConversation which references the
conversation ID of the currently open chat. when this value is set, an open
conversation message is sent to the server, and when it is cleared, a close
conversation message is sent to server.

the value is also set and cleared on the onSocketConnected and onDestroy
callbacks respectively