Use ID and timestamp to update DLed messages
To download messages you must now pass the message's conversation ID and sorting
timestamp along with the message ID and sender ID. This change was made so that
you can get back this data on the onFileDownloadFinished callback for easier
message searching.
On MainActivity, when a file is downloaded completely, the conversation ID is
used to check if the active conversation is where the message belongs. If it
does, then let the Chatfragment update it. Otherwise look for in the chatMap and
update.

ChatMap now holds Lists instead of collections.