replace onMessageBatchReady with onSyncComplete
The onMessageBatchReady callback has been replaced with the
onSyncComplete callback. This new callback receivers a list of
notifications and a list of deletes, as well as a list of messages.

MOKNotification and MOKDelete are new classes that represent the
notification and the delete.

The getBatch function of HTTPSync now returns a SyncResponse object instead of a
list of messages.