after sync return map of messages instead of list
new class SyncData has a HashMap where the key is a conversationId and
the value is a list of messages for that conversation. onSyncComplete
now uses this type instead of SyncResponse, to update conversations more
easily.

onSyncComplete now iterates over this map to add messages to its
respective conversation. A HashMap of ids and transactions is generated
to update the conersations fragment efficiently. Then it calls the
new AsyncTask UpdateConversationsTask that updates all of these
conversations in the database.

Fixed issue with messages not being decrypted properly in the
getConversations method of UserManager.

use UI Kit commit be42c64d75