delete messages and update conversation after sync
onSyncComplete now deletes unsent messages and updates conversation
correctly.

New AsyncTask DeleteMessagesTask deletes messages from the database and
updates the conversations that deleted unread messages. Then. it returns
to the main thread the conversation transactions that should be applied
to the UI.

SyncData now has a hashMap of deletes so that they can be procesed per
conversation.

getMessagePage is now executed in parallel because symc can block the
AsyncTask thread for a long time.

Use UI Kit commit b4ab456082