Add new column to ConversationItem: lastOpen
ConversationItem has a new column that keeps track of a timestamp of the
last time the user opened a conversation. This is useful to calculate
the number of unread messages during sync.

New class ConversationManager has logic for dealing with conversations,
eventually all MainActivity's code that deals with conversation will be
in this class.

The code that deals with deletes has been commented out because it can
potentially throw lots of NPEs