Sync using ConversationItem's lastOpen
Sync Response handling has been reemplemented so that it can use the
conversationItem's lastOpen value to compute the actual number of new
messages after receiving the sync response.

ConversationManager class has been deleted. MainActivity has a new field
activeConversationItem to keep a reference of the ConversationItem
object of the active chat.

Messages are no longer loaded in pages. There is no need to keep track
of the current page.  The Database can figure out the neccesary offset
using the number of loaded messages as input.

AsyncDBHandler has new tasks FindConversationsTask which searches fo a
list of Conversations in the database using the IDs and
GetConversationPageTask which loads a complete "page" of conversations.

MonkeyKit uses the AsyncTask SyncDatabaseTask to make the sync HTTP
Request in background.

Use UI Kit commit c8788439b7.