set correct badge on new conversations
When creating new conversations after the response of getUserInfo() or
getGroupInfo() the local database is queried to retrieve the
totalNewMessages and secondary text. The badge is now displayed more
accurately.

The acknowledge of read messages is now processed correctly. instead of
updating messages in the database, just update the lastRead value of the
MonkeyAdapter.

Use UI Kit 9f448d3c5b