Show a "connecting" message if sync gets delayed
new class SyncStatus has a handler and a weak reference to
MonkeyFragmentManager to wait for a timeout. When the timeout occurs, if
sync response has not arrived, show a connecting message. In this
scenario, once the sync response arrives show a "connected" message.

If sync is not delayed, no message is ever shown.