Use Robolectric 3.2
New Robolectric release honors @Config(manifest="___") which means that
we no longer need to specify an app theme in the SDK manifest.
With this change is easier to add the as a dependency to an
existing project.

UI Kit commit c9e8e39447 also uses this Robolectric version. This solves
issue #87