Fix issue #89
The reason why conversations where not being updated is because the
ConversationTransaction objects that updated the UI were not setting the
conversation status as 'sentMessageRead'. The functions that create this
transactions have now been moved to TransactionCreator class.

If the conversation is updated from the ACK of a sent message then
fromSentMessage() is used. This function receives a 'deliveredAndRead'
parameter to toggle between 'deliveredMessage' and 'sentMessageRead'.

If the conversation is updated after the contact opens the conversation
the fromContactOpenedConversation() is used. This function will always
set the status as 'sentMessageRead' if the previous status was
'deliveredMessage'.

The models package has been reestructured for easier navigation.
AsyncTasks now have their own package.`