Fix bug parsing last_seen in open conversation
open conversation response returns a last_seen value as string if it is
a normal conversation, but if it is a group, it returns it as an
object. The quick solution for this problem is to iterate the keys of
the object and get the minimum value. This is the solution we will use
until we support multiple last_seen values for conversations. This
solves issue #93 for now.