fix issue new messages not displayed after sync
In the onSyncComplete() callback of MainActivity all conversations that
had new messages would remove the MessagesList object from memory, so
that a new up to date object would be retrieved from the database. The
problem was that the active conversation would also remove its messages,
so the MonkeyChatFragment would have a different MessagesList object
from the one in the Activity's state.

To fix this, a new class "SyncState" was created. this performs all the
updates needed in the activity's state after sync and returns the number
of new messages in the active conversation. With this number,
MainActivity can decide wether MonkeyChatFragment needs an update.
SyncState does not clear the MessagesList object of the active
conversation.

MainActivity's state is now kept in a single class: "ChatState".
ChatDataFragment keeps a reference to this object. MainActivity now has
a new function to access state: getState(), instead of accessing
attributes of ChatDataFragment. This is a big refactor.

Use UI Kit commit 28024310ca. This guarantees that when the fragment is
stopped, the list object does not keep a reference to the fragment. This
applies to MonkeyChatFragment and MonkeyConversationsFragment.