New methods to deal with pending msgs Json objects
Messages that the socket cannot send are stored as JSON in a text file,
so that they are resent the next time the socket connects. If the
service restarts this file has to be parsed again. Some issues were
found while parsing this file

ResponseParser has been renamed to MonkeyJson. This class has new
methods parsePendingMsgsFromFile() and sanitizePendingMsgsForFile()
which prevent errors while serializing/deserializing these JsonObjects.