new delegate: ConversationOpenDelegate
This delegate takes away the functions onConversationOpenResponse and
onContactOpenMyConversation from ConversationDelegate. The reason is
that these two can be triggered randomly at any point, while the rest of
the methods in ConversationDelegate are only trigged after user action.

Added logs to debug AsyncConnSocket. These have to be eventaully removed
when we can solve some issues.