Persist pending messages as soon as they are sent
MonkeyKitSocketService has a new attribute: pendingMessageStore which
replaces the pendingMessages list. This class triggers the task that
stores the pending messages to a file as soon as they are sent instead
of waiting for onDestroy() to do it. Since onDestroy is not guaranteed
to be called, this is a safer approach.