Fix addmember notification error
When user is added to group, two notifications are received:
groupAddMember and groupCreate. Group create notification is now parsed
correctly to create the actual group convetsation. New class
NotificationMessages handles this logic in a more testable way.

DatabaseHandler is no longer a static class, you have to create an
instance to use it. this makes it easier to mock methods.